Refactor: 중복되는img 태그 분리
