+++ src/component/home/Section03Post.jsx
@@ -1,12 +1,10 @@
 import styled from 'styled-components';
+import ImgTag from '../common/ImgTag';
 
 const Section03Post = ({ title, address, imgurl }) => {
-  const noImgURL = '/images/board/no-img.jpg';
-  const handleImageError = e => (e.target.src = noImgURL);
-
   return (
     <PostContainer>
-      <img src={imgurl || noImgURL} onError={handleImageError} alt="dog Image" />
+      <ImgTag img={imgurl} height="70%" />
       <TextLabelContainer>
         <h1>{title}</h1>
         <h5>{address}</h5>
@@ -24,12 +22,6 @@ const PostContainer = styled.div`
   background-color: white;
 
   overflow: hidden;
-
-  img {
-    width: 100%;
-    height: 70%;
-    object-fit: cover;
-  }
 `;
 
 const TextLabelContainer = styled.div`
